requirements.txt
google-genai>=0.8.0
openai>=1.0.0
pystray>=0.19.0
Pillow>=10.0.0
